moved app to src
